WIP
WIP

WIP

feat(listall): Add copy to clipboard functionality

Use clipboard.js library to enable copy to clipboard functionality.
Add a button to the right of each voucher code's explanation which allows
the code being displayed to be copied to clipboard when clicked.
The header of coloumn where the buttons are placed shows a notification when
the code is copied successfully. No notification if the copy fails at the moment.

Miscellaneous:

Modify how the domain / host name is sent to the server.
There is a function called function getCompanyFromHost(host) in insert.js which
attempts to extract the company name from the FQDN. This allows for more relevant
voucher codes to be displayed in most circumstances.
This functionality should be moved to the backend and will be done so with time.
Exceptions are to be made for those sites where the FQDN displays more and
relevant vouchers than the company name.